+++ routes/newProduct.js
@@ -4,19 +4,9 @@ const axios = require("axios");
 const fs = require("fs");
 const validatePreferences = require("../utilities/handlePreferences");
 const addCurrentProducts = require("../utilities/handleCurrentData");
+const formateProductName = require("../utilities/formateProductName");
 const baseAPI = process.env.API_URl;
 const serpapiKey = process.env.API_KEY;
-
-const formateProductName = (productName) => {
-  const productNameArray = productName.split(" ");
-  let productNameConcate = "";
-  for (let index = 0; index < productNameArray.length; index++) {
-    productNameConcate = productNameConcate
-      ? `${productNameConcate}+${productNameArray[index]}`
-      : productNameArray[index];
-  }
-  return productNameConcate;
-};
 
 router.get("/products", (req, res, next) => {
   // ADDING PREFERENCES TO DATABASE
